+++ lib/model/client/client_account_update_context.dart
@@ -1,41 +1,39 @@
-import '../../enums/client/client_sex.dart';
+import 'client_sex.dart';
 import 'client_agreements.dart';
 
 /// The `ClientAccountUpdateContext` class represents the context for updating a client's account
 /// information, including personal details, agreements, and custom attributes.
 class ClientAccountUpdateContext {
   String? email;
-  String? password;
+  String? customId;
+  String? uuid;
   String? firstName;
   String? lastName;
   ClientSex? sex;
   String? phone;
   String? company;
   String? address;
+  String? province;
   String? city;
   String? zipcode;
   String? countrycode;
-  String? province;
-  String? uuid;
-  String? customId;
   ClientAgreements? agreements;
   Map<String, Object>? attributes;
 
   ClientAccountUpdateContext(
       {this.email,
-      this.password,
+      this.customId,
+      this.uuid,
       this.firstName,
       this.lastName,
       this.sex,
       this.phone,
       this.company,
       this.address,
+      this.province,
       this.city,
       this.zipcode,
       this.countrycode,
-      this.province,
-      this.uuid,
-      this.customId,
       this.agreements,
       this.attributes});
 
@@ -44,19 +42,18 @@ class ClientAccountUpdateContext {
   /// the values correspond to the property values.
   Map asMap() => {
         'email': email,
-        'password': password,
+        'customId': customId,
+        'uuid': uuid,
         'firstName': firstName,
         'lastName': lastName,
         'sex': sex?.getSexAsString(),
         'phone': phone,
         'company': company,
         'address': address,
+        'province': province,
         'city': city,
         'zipcode': zipcode,
         'countrycode': countrycode,
-        'province': province,
-        'uuid': uuid,
-        'customId': customId,
         'agreements': agreements?.asMap(),
         'attributes': attributes
       };